move MockValidPassword and MockHashPassword into database package
